Fix mobile sdk functions - GetEthClient - Logger
