+++ zcncore/networkworker_mobile.go
@@ -47,15 +47,15 @@ func updateNetworkDetailsWorker(ctx context.Context) {
 	for {
 		select {
 		case <-ctx.Done():
-			Logger.Info("Network stopped by user")
+			logging.Info("Network stopped by user")
 			return
 		case <-ticker.C:
 			err := UpdateNetworkDetails()
 			if err != nil {
-				Logger.Error("Update network detail worker fail", zap.Error(err))
+				logging.Error("Update network detail worker fail", zap.Error(err))
 				return
 			}
-			Logger.Info("Successfully updated network details")
+			logging.Info("Successfully updated network details")
 			return
 		}
 	}
@@ -64,7 +64,7 @@ func updateNetworkDetailsWorker(ctx context.Context) {
 func UpdateNetworkDetails() error {
 	networkDetails, err := GetNetworkDetails()
 	if err != nil {
-		Logger.Error("Failed to update network details ", zap.Error(err))
+		logging.Error("Failed to update network details ", zap.Error(err))
 		return err
 	}
